# core/memory/working/projects/rate-my-set/IN/better-base-dev/openapi/generated/schema.yml
openapi: 3.0.3
info:
  title: Better Base API
  version: 0.1.0
  description: Documentation of API endpoints of Better Base
paths:
...
  /api/invitations/{id}/accept:
    post:
      operationId: api_invitations_accept_create
      parameters:
      - in: path
        name: id
        schema:
          type: integer
        description: A unique integer value identifying this Invitation.
        required: true
      tags:
      - api
      requestBody:
        content:
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/InvitationAccept'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/InvitationAccept'
          application/json:
            schema:
              $ref: '#/components/schemas/InvitationAccept'
      security:
      - cookieAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InvitationAccept'
          description: ''
  /api/invitations/{id}/decline:
    post:
      operationId: api_invitations_decline_create
      parameters:
      - in: path
        name: id
        schema:
          type: integer
        description: A unique integer value identifying this Invitation.
        required: true
      tags:
      - api
      requestBody:
        content:
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/InvitationDecline'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/InvitationDecline'
          application/json:
            schema:
              $ref: '#/components/schemas/InvitationDecline'
      security:
      - cookieAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InvitationDecline'
          description: ''
  /api/invitations/{id}/resend:
    post:
      operationId: api_invitations_resend_create
      parameters:
      - in: path
        name: id
        schema:
          type: integer
        description: A unique integer value identifying this Invitation.
        required: true
      tags:
      - api
      requestBody:
        content:
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/InvitationResend'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/InvitationResend'
          application/json:
            schema:
              $ref: '#/components/schemas/InvitationResend'
      security:
      - cookieAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InvitationResend'
          description: ''
...
components:
  schemas:
...
    AccountReadOnly:
      type: object
      properties:
        id:
          type: integer
          readOnly: true
        account_type:
          allOf:
          - $ref: '#/components/schemas/AccountTypeEnum'
          readOnly: true
        account_type_display:
          type: string
          readOnly: true
        name:
          type: string
          readOnly: true
        uploaded_profile_image:
          type: string
          format: uri
          readOnly: true
        fallback_name:
          type: string
          readOnly: true
        display_name:
          type: string
          readOnly: true
        created:
          type: string
          format: date-time
          readOnly: true
      required:
      - account_type
      - account_type_display
      - created
      - display_name
      - fallback_name
      - id
      - name
      - uploaded_profile_image
...
    AccountTypeEnum:
      enum:
      - personal
      - team
      type: string
      description: |-
        * `personal` - Personal
        * `team` - Team
...
    DeliveryMethodEnum:
      enum:
      - email
      type: string
      description: '* `email` - Email'
    InvitationAccept:
      type: object
      properties:
        id:
          type: integer
          readOnly: true
        account:
          allOf:
          - $ref: '#/components/schemas/AccountReadOnly'
          readOnly: true
        invited_by:
          allOf:
          - $ref: '#/components/schemas/UserReadOnly'
          readOnly: true
        email:
          type: string
          format: email
          readOnly: true
        name:
          type: string
          readOnly: true
        role:
          allOf:
          - $ref: '#/components/schemas/RoleEnum'
          readOnly: true
        role_display:
          type: string
          readOnly: true
        user:
          allOf:
          - $ref: '#/components/schemas/UserReadOnly'
          readOnly: true
        accepted_at:
          type: string
          format: date-time
          readOnly: true
          nullable: true
        expires_at:
          type: string
          format: date-time
          readOnly: true
        delivery_method:
          allOf:
          - $ref: '#/components/schemas/DeliveryMethodEnum'
          readOnly: true
        last_sent_at:
          type: string
          format: date-time
          readOnly: true
          nullable: true
        created:
          type: string
          format: date-time
          readOnly: true
        is_accepted:
          type: boolean
          readOnly: true
        is_declined:
          type: boolean
          readOnly: true
        is_expired:
          type: boolean
          readOnly: true
        is_past_follow_window:
          type: boolean
          readOnly: true
        status:
          type: string
          readOnly: true
        status_display:
          type: string
          readOnly: true
        team_display_name:
          type: string
          readOnly: true
        is_using_fallback_team_display_name:
          type: boolean
          readOnly: true
        headline:
          type: string
          readOnly: true
      required:
      - accepted_at
      - account
      - created
      - delivery_method
      - email
      - expires_at
      - headline
      - id
      - invited_by
      - is_accepted
      - is_declined
      - is_expired
      - is_past_follow_window
      - is_using_fallback_team_display_name
      - last_sent_at
      - name
      - role
      - role_display
      - status
      - status_display
      - team_display_name
      - user
...
    InvitationDecline:
      type: object
      properties:
        id:
          type: integer
          readOnly: true
        account:
          allOf:
          - $ref: '#/components/schemas/AccountReadOnly'
          readOnly: true
        invited_by:
          allOf:
          - $ref: '#/components/schemas/UserReadOnly'
          readOnly: true
        email:
          type: string
          format: email
          readOnly: true
        name:
          type: string
          readOnly: true
        role:
          allOf:
          - $ref: '#/components/schemas/RoleEnum'
          readOnly: true
        role_display:
          type: string
          readOnly: true
        user:
          allOf:
          - $ref: '#/components/schemas/UserReadOnly'
          readOnly: true
        accepted_at:
          type: string
          format: date-time
          readOnly: true
          nullable: true
        expires_at:
          type: string
          format: date-time
          readOnly: true
        delivery_method:
          allOf:
          - $ref: '#/components/schemas/DeliveryMethodEnum'
          readOnly: true
        last_sent_at:
          type: string
          format: date-time
          readOnly: true
          nullable: true
        created:
          type: string
          format: date-time
          readOnly: true
        is_accepted:
          type: boolean
          readOnly: true
        is_declined:
          type: boolean
          readOnly: true
        is_expired:
          type: boolean
          readOnly: true
        is_past_follow_window:
          type: boolean
          readOnly: true
        status:
          type: string
          readOnly: true
        status_display:
          type: string
          readOnly: true
        team_display_name:
          type: string
          readOnly: true
        is_using_fallback_team_display_name:
          type: boolean
          readOnly: true
        headline:
          type: string
          readOnly: true
      required:
      - accepted_at
      - account
      - created
      - delivery_method
      - email
      - expires_at
      - headline
      - id
      - invited_by
      - is_accepted
      - is_declined
      - is_expired
      - is_past_follow_window
      - is_using_fallback_team_display_name
      - last_sent_at
      - name
      - role
      - role_display
      - status
      - status_display
      - team_display_name
      - user
...
    InvitationResend:
      type: object
      properties:
        id:
          type: integer
          readOnly: true
        account:
          allOf:
          - $ref: '#/components/schemas/AccountReadOnly'
          readOnly: true
        invited_by:
          allOf:
          - $ref: '#/components/schemas/UserReadOnly'
          readOnly: true
        email:
          type: string
          format: email
          readOnly: true
        name:
          type: string
          readOnly: true
        role:
          allOf:
          - $ref: '#/components/schemas/RoleEnum'
          readOnly: true
        role_display:
          type: string
          readOnly: true
        user:
          allOf:
          - $ref: '#/components/schemas/UserReadOnly'
          readOnly: true
        accepted_at:
          type: string
          format: date-time
          readOnly: true
          nullable: true
        expires_at:
          type: string
          format: date-time
          readOnly: true
        delivery_method:
          allOf:
          - $ref: '#/components/schemas/DeliveryMethodEnum'
          readOnly: true
        last_sent_at:
          type: string
          format: date-time
          readOnly: true
          nullable: true
        created:
          type: string
          format: date-time
          readOnly: true
        is_accepted:
          type: boolean
          readOnly: true
        is_declined:
          type: boolean
          readOnly: true
        is_expired:
          type: boolean
          readOnly: true
        is_past_follow_window:
          type: boolean
          readOnly: true
        status:
          type: string
          readOnly: true
        status_display:
          type: string
          readOnly: true
        team_display_name:
          type: string
          readOnly: true
        is_using_fallback_team_display_name:
          type: boolean
          readOnly: true
        headline:
          type: string
          readOnly: true
      required:
      - accepted_at
      - account
      - created
      - delivery_method
      - email
      - expires_at
      - headline
      - id
      - invited_by
      - is_accepted
      - is_declined
      - is_expired
      - is_past_follow_window
      - is_using_fallback_team_display_name
      - last_sent_at
      - name
      - role
      - role_display
      - status
      - status_display
      - team_display_name
      - user
...
    IsAuthenticatedEnum:
      enum:
      - true
      type: boolean
...
    RoleEnum:
      enum:
      - member
      - owner
      type: string
      description: |-
        * `member` - Member
        * `owner` - Owner
...
    UserReadOnly:
      type: object
      properties:
        id:
          type: integer
          readOnly: true
        email:
          type: string
          format: email
          readOnly: true
        email_is_verified:
          type: boolean
          readOnly: true
          title: Email Is Verified?
          description: Has this User's email address been verified?
        email_verified_as_of:
          type: string
          format: date-time
          readOnly: true
          nullable: true
          description: When was this User's email address verified?
        name:
          type: string
          readOnly: true
        uploaded_profile_image:
          type: string
          format: uri
          readOnly: true
        is_authenticated:
          allOf:
          - $ref: '#/components/schemas/IsAuthenticatedEnum'
          readOnly: true
        is_active:
          type: boolean
          readOnly: true
          title: Is Active?
          description: Designates whether this user should be treated as active. Potentially
            unselect this instead of deleting accounts.
        is_staff:
          type: boolean
          readOnly: true
          title: Is Staff?
          description: Designates whether the user can log into this admin site.
        is_superuser:
          type: boolean
          readOnly: true
          title: Superuser status
          description: Designates that this user has all permissions without explicitly
            assigning them.
        last_login:
          type: string
          format: date-time
          readOnly: true
          nullable: true
        created:
          type: string
          format: date-time
          readOnly: true
      required:
      - created
      - email
      - email_is_verified
      - email_verified_as_of
      - id
      - is_active
      - is_authenticated
      - is_staff
      - is_superuser
      - last_login
      - name
      - uploaded_profile_image
...
  securitySchemes:
    CsrfExemptSessionAuthentication:
      type: apiKey
      in: cookie
      name: sessionid
    cookieAuth:
      type: apiKey
      in: cookie
      name: sessionid